научил выводить цифры после запятой

+++ src/operations.java
@@ -1,13 +1,13 @@
 public class operations {
     private int firstNumber = 0;
     private int secondNumber = 0;
-    private int result = 0;
+    private double result = 0;
 
 
     public operations (int firstNumber, String oper, int secondNumber) {
     switch (oper) {
         case ("/"):
-            result = firstNumber / secondNumber;
+            result = (double) firstNumber / secondNumber;
             break;
         case ("*"):
             result = firstNumber * secondNumber;
@@ -26,7 +26,7 @@ public class operations {
 
           if (calcThis.flag) {
               if (result >= 1) {
-                  return "Ответ " + arabToRim.convertIntegerToRoman(result);
+                  return "Ответ " + arabToRim.convertIntegerToRoman((int) result);
               } else {return "Число не отображается римскими цифрами";
               }
           } else {
